+++ src/home.js
@@ -1,6 +1,12 @@
 const renderHomePage = (() => {
     const contentContainer = document.querySelector('#content');
-    contentContainer.innerHTML=`default page`
-
+    contentContainer.innerHTML=`home`
+    
+    const homeLink = document.querySelector('#home-link')
+    homeLink.addEventListener("click", e => {
+      contentContainer.innerHTML = `home`
+    })
   })();
   export { renderHomePage };
+
+
