+++ src/menu.js
@@ -1,5 +1,10 @@
-const menuSelector = document.querySelector('#menu-link');
-const renderMenuPage = (() => {
+const renderMenu = (() => {
+    const menuLink = document.querySelector('#menu-link');
+    const body = document.querySelector('#content');
+    
+    menuLink.addEventListener("click", e => {
+      body.innerHTML = `menu`
+    })
 
-});
-export { renderMenuPage };
+  })();
+  export { renderMenu };
